Just in time compilation for riscv32 architecture

Continuation of:
- #1881
- #1829
- #1891
- #1838
- #1900
- #1901
- #1903
- #1981

These changes are made under both the "Apache 2.0" and the "GNU Lesser General
Public License 2.1 or later" license terms (dual license).

SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later

+++ src/libAtomVM/opcodesswitch.h
@@ -7738,7 +7738,7 @@ handle_error:
         }
 
 terminate_context:
-        TRACE("-- Code execution finished for %i--\n", ctx->process_id);
+        TRACE("-- Code execution finished for %i--\n", (int) ctx->process_id);
         GlobalContext *global = ctx->global;
         if (ctx->leader) {
             scheduler_stop_all(global);
